+++ res/images/selectAlgo.gif
